// src/types/endpoints/common/NFLTeam.ts
enum NFLTeam {
  arizonaCardinals = "arizonaCardinals",
  atlantaFalcons = "atlantaFalcons",
  carolinaPanthers = "carolinaPanthers",
  chicagoBears = "chicagoBears",
  dallasCowboys = "dallasCowboys",
  detroitLions = "detroitLions",
  greenBayPackers = "greenBayPackers",
  losAngelesRams = "losAngelesRams",
  minnesotaVikings = "minnesotaVikings",
  newOrleansSaints = "newOrleansSaints",
  newYorkGiants = "newYorkGiants",
  philadelphiaEagles = "philadelphiaEagles",
  sanFrancisco49ers = "sanFrancisco49ers",
  seattleSeahawks = "seattleSeahawks",
  tampaBayBuccaneers = "tampaBayBuccaneers",
  washingtonCommanders = "washingtonCommanders",
  baltimoreRavens = "baltimoreRavens",
  buffaloBills = "buffaloBills",
  cincinnatiBengals = "cincinnatiBengals",
  clevelandBrowns = "clevelandBrowns",
  denverBroncos = "denverBroncos",
  houstonTexans = "houstonTexans",
  indianapolisColts = "indianapolisColts",
  jacksonvilleJaguars = "jacksonvilleJaguars",
  kansasCityCheifs = "kansasCityCheifs",
  lasVegasRaiders = "lasVegasRaiders",
  losAngelesChargers = "losAngelesCharges",
  miamiDolphins = "miamiDolphins",
  newEnglandPatriots = "newEnglandPatriots",
  newYorkJets = "newYorkJets",
  pittsburghSteelers = "pittsburghSteelers",
  tennesseeTitans = "tennesseeTitans"
}

export default NFLTeam;